Updated to version 10.6.2 (code 312)
* In build.gradle, updated Crashlytics version.

* In app/build.gradle, updated package versions.

* In activealarm/NacActiveAlarmService.java, specified casting to long
  when calculating wakelock timeout.

* In activealarm/NacWakeupProcess.java, changed lambda expressions to
  method references.

* In activealarm/NacWakeupProcess.java, removed unused imports.

* In activealarm/NacWakeupProcess.java, removed print statements.

* In card/NacCardHolder.java, fixed issue where the OnCardUpdateListener
  was still being called after the SeekBar called onStopTrackingTouch().
  Added a check in onStopTrackingTouch() to make sure the alarm is not
  in use before calling the listener.

* In card/NacCardHolder.java, updated calls to NacAlarm that were
  passing in a NacSharedPreferences object to check the snooze count.
  The logic of the code probably never got here, but better to be safe.

* In main/NacMainActivity.java, added check for API 31 before doing
  anything in onPermissionRequestDone().

* In media/NacAudioAttributes.java, removed unused imports.

* In media/NacAudioManager.java, removed unused imports.

* In media/NacMediaPlayer.java, removed unused imports.

* In nfc/NacNfc.java, removed a print statement.

* In
  permission/NacScheduleExactAlarmPermissionChangedBroadcastReceiver.java,
  added a check for API 31 before doing anything in onReceive().

* In permission/NacScheduleExactAlarmPermissionDialog.java, removed
  unused print statements.

* In scheduler/NacScheduler.java, removed unused imports.

* In tts/NacTextToSpeech.java, removed print statements.

* In whatsnew/NacWhatsNewDialog.java, removed unused print statements.

* In Spanish arrays.xml, added "Gradually increase volume" to the audio
  options.

* In Spanish strings.xml, added everything for the "What's new" dialog.

* In Spanish strings.xml, added strings for "Gradually increase volume"
  and "Restrict volume".

* In English strings.xml, tweaked the wording of the "What's new" dialog
  and "Gradually increase volume".

+++ app/src/main/java/com/nfcalarmclock/activealarm/NacActiveAlarmService.java
@@ -535,7 +535,7 @@ public class NacActiveAlarmService
 		PowerManager pm = (PowerManager) getSystemService(
 			Context.POWER_SERVICE);
 		this.mWakeLock = pm.newWakeLock(PowerManager.PARTIAL_WAKE_LOCK, tag);
-		long timeout = shared.getAutoDismissTime() * 60 * 1000;
+		long timeout = (long) shared.getAutoDismissTime() * 60L * 1000L;
 
 		this.mWakeLock.acquire(timeout);
 	}
